feat: 통계용 팝업리스트 PopupController-> StatisticsController로 분리
